+++ src/main/java/com/beehyv/case_study/entities/Cart.java
@@ -7,15 +7,15 @@ import java.util.List;
 public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int cartId;
+    private long cartId;
     @OneToMany
     private List<CartItem> products;
 
-    public int getCartId() {
+    public long getCartId() {
         return cartId;
     }
 
-    public void setCartId(int cartId) {
+    public void setCartId(long cartId) {
         this.cartId = cartId;
     }
