+++ src/main/java/com/beehyv/case_study/entities/Product.java
@@ -13,18 +13,18 @@ import java.util.Objects;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int productId;
+    private long productId;
     private String name;
     private int price;
     private String details;
     private String category;
     private String subcategory;
 
-    public int getProductId() {
+    public long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(long productId) {
         this.productId = productId;
     }
