+++ src/main/java/com/beehyv/case_study/repositories/MyUserCredentialsRepo.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MyUserCredentialsRepo extends JpaRepository<MyUserCredentials, Integer> {
+public interface MyUserCredentialsRepo extends JpaRepository<MyUserCredentials, Long> {
     Optional<MyUserCredentials> getByEmail(String email);
-
     Boolean existsByEmail(String email);
 }
